refactor: add useGlobalKeyboardShortcut
